Use config dialog to set sample rate
As a consequence of this, LengthSamples and FadeoutSamples have been
scrapped, as these are dependent on the sample rate at the
time. Wherever these properties were used, they are now being
calculated dynamically based on what the plugin state says the sample
rate is. However, there is room to optimise this (e.g. calculate this
once and store it in a variable, then reuse this variable in the rest
of the function).