Fix bug in end-of-song logic that caused copying of partial samples
Previously, we directly compared `to_copy` (bytes requested by deadbeef)
with `remaining_samples` (samples remaining in the current song), and
would set `to_copy = remaining_samples` when `to_copy >
`remaining_samples`. Since one sample is equal to 4 bytes, not 1 byte,
this would cause a partial sample to be copied if `remaining_samples`
is not a multiple of 4, which would cause conversions to fail and
could cause noise to be played at the start of the next song.

Now, `remaining_samples` is converted to `remaining_bytes =
remaining_samples * 4`, and this quantity is used for detecting the
end of the song.